backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pyjwt==2.10.1
